Rename `pspamm.pspamm` to `pspamm.cli`
