+++ pspamm.py
@@ -1,4 +1,5 @@
 #!/usr/bin/env python3
+
 if __name__=='__main__':
-    import pspamm.pspamm
-    pspamm.pspamm.main()
+    import pspamm.cli
+    pspamm.cli.main()
